add username validation to server

+++ randomideas-api/routes/ideas.js
@@ -13,7 +13,7 @@ router.get('/', async (req, res) => {
   }
 });
 
-// Get an idea
+// Get single idea
 router.get('/:id', async (req, res) => {
   try {
     const idea = await Idea.findById(req.params.id);
@@ -24,7 +24,7 @@ router.get('/:id', async (req, res) => {
   }
 });
 
-// Add idea
+// Add an idea
 router.post('/', async (req, res) => {
   const idea = new Idea({
     text: req.body.text,
@@ -33,39 +33,61 @@ router.post('/', async (req, res) => {
   });
 
   try {
-    const saveIdea = await idea.save();
-    res.json({ success: true, data: saveIdea });
+    const savedIdea = await idea.save();
+    res.json({ success: true, data: savedIdea });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
   }
 });
 
-// update idea
+// Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to update this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
   }
 });
 
-// delete idea
+// Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to delete this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
